requirements.txt
-i https://pypi.org/simple
aenum==3.1.15
aiosqlite==0.17.0; python_version >= '3.6'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.4.0; python_version >= '3.8'
asyncmy==0.2.9
certifi==2024.7.4; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; sys_platform == 'win32'
dbf==0.99.9
dnspython==2.6.1; python_version >= '3.8'
docopt==0.6.2
email-validator==2.2.0
exceptiongroup==1.2.2; python_version < '3.11'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.4; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
iso8601==1.1.0; python_version < '4.0' and python_full_version >= '3.6.2'
jinja2==3.1.4; python_version >= '3.7'
loguru==0.7.2; python_version >= '3.5'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
num2words==0.5.13
numpy==2.0.0; python_version < '3.11'
orjson==3.10.6; python_version >= '3.8'
pandas==2.2.2; python_version >= '3.9'
pydantic[email]==2.8.2; python_version >= '3.8'
pydantic-core==2.20.1; python_version >= '3.8'
pydantic-settings==2.3.4; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pypika-tortoise==0.1.6; python_version >= '3.7' and python_version < '4.0'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.9; python_version >= '3.8'
pytz==2024.1
pyyaml==6.0.1
requests==2.32.3; python_version >= '3.8'
rich==13.7.1; python_full_version >= '3.7.0'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
tortoise-orm[asyncmy]==0.21.4; python_version >= '3.8' and python_version < '4.0'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2024.1; python_version >= '2'
ujson==5.10.0; python_version >= '3.8'
urllib3==2.2.2; python_version >= '3.8'
uvicorn[standard]==0.30.1; python_version >= '3.8'
watchfiles==0.22.0
websockets==12.0
win32-setctime==1.1.0; sys_platform == 'win32'
